Allow user to define meta tags
When the `MetaTagsPart` is attached to a content type, content managers
will be able to define title, description and an image. The content for
these fields will be used to generate meta tags for generic (title &
description) as well as open graph & twitter meta tags.

Used https://metatags.io/ to determine the meta tags that are a priority.