Add Azure Pipelines build scripts
Add build scripts for working with Azure DevOps Pipeline builds.

The azure-pipelines-prerelease will build a package with a prerelease
suffix of CI and the current date/time. This allows continuous integration
for easy testing and development of packages along side a site which uses
them. The version number itself is based on the assembly version number.

The azure-pipelines-release build is exactly the same except it simply
uses the assembly major.minor.patch to create a stable version for the
current version number for deployment to a Nuget feed.