Change location of `MetaTagsPart` display
Was being pushed to "Header" zone, however it's not always certain that
a content type will have a "Header" zone. The shape only adds to the
meta title of the page and doesn't render anything that is visible to
the user. Using "Content" zone instead as it's more likely to exist when
rendering a content item.

Related to #18.