Fix migration issue caused when migrating meta tags
Noticed an issue when upgrading a site recently that would cause an
exception because the `Autouroute` migration has not been able to run
before a query is made to get content items that need their meta data
updated. Solution is to add a dependency to `OrchardCore.Autoroute` to
ensure that migration is executed before the SEO migrations.